address warnings; move API into own sub-module
