+++ python/casm/casm/project/__init__.py
@@ -4,13 +4,11 @@ from casm.project.project import project_path, ClexDescription, ProjectSettings,
 from casm.project.selection import Selection
 from casm.project.query import query
 from casm.project.io import write_eci
-from casm.project.api import API
 __all__ = [
-  'API',
   'project_path',
-  'ClexDescription', 
-  'ProjectSettings', 
-  'DirectoryStructure', 
+  'ClexDescription',
+  'ProjectSettings',
+  'DirectoryStructure',
   'Project',
   'Prim',
   'Selection',
