+++ python/casm/casm/project/project.py
@@ -12,7 +12,7 @@ import numpy as np
 import six
 
 from casm.project import syminfo
-from casm.project.api import API
+from casm.api import API
 
 def project_path(dir=None):
     """
@@ -43,28 +43,28 @@ def project_path(dir=None):
 class ClexDescription(object):
     """
     Settings for a cluster expansion
-    
+
     Attributes
     ----------
-      
+
       name: str
         Cluster expansion name
-      
+
       property: str
         Name of the property being cluster expanded
-      
+
       calctype: str
         Calctype name
-      
+
       ref: str
         Reference state name
-      
+
       bset: str
         Basis set
-      
+
       eci: str
         ECI set name
-      
+
     """
     def __init__(self, name, property, calctype, ref, bset, eci):
       self.name = name
@@ -78,18 +78,18 @@ class ClexDescription(object):
 class ProjectSettings(object):
     """
     Settings for a CASM project
-    
+
     Attributes:
       path: Path to CASM project
       data: Dict storing contents of project_settings.json
-    
+
     """
     def __init__(self, path=None):
         """
         Construct a CASM ProjectSettings representation.
 
         Args:
-            path: path to CASM project (Default=None, uses project containing current directory). 
+            path: path to CASM project (Default=None, uses project containing current directory).
 
         """
         if project_path(path) is None:
@@ -105,29 +105,29 @@ class ProjectSettings(object):
         self._default_clex = ClexDescription(d["name"], d["property"], d["calctype"], d["ref"], d["bset"], d["eci"])
 
         self._clex = [
-          ClexDescription(d[1]["name"], d[1]["property"], d[1]["calctype"], d[1]["ref"], d[1]["bset"], d[1]["eci"]) 
+          ClexDescription(d[1]["name"], d[1]["property"], d[1]["calctype"], d[1]["ref"], d[1]["bset"], d[1]["eci"])
           for d in six.iteritems(self.data["cluster_expansions"])
         ]
-        
+
         d = self.data["cluster_expansions"].get("formation_energy", None)
         self._formation_energy_clex = None
         if d is not None:
           self._formation_energy_clex = ClexDescription(d["name"], d["property"], d["calctype"], d["ref"], d["bset"], d["eci"])
-    
+
     # -- Accessors --
-    
+
     @property
     def cluster_expansions(self):
         return self._clex
-    
+
     @property
     def default_clex(self):
         return self._default_clex
-    
+
     @property
     def formation_energy_clex(self):
         return self._formation_energy_clex
-    
+
 
 class DirectoryStructure(object):
     """Standard file and directory locations for a CASM project"""
@@ -136,7 +136,7 @@ class DirectoryStructure(object):
         Construct a CASM Project DirectoryStructure representation.
 
         Args:
-            path: path to CASM project (Default=None, uses project containing current directory). 
+            path: path to CASM project (Default=None, uses project containing current directory).
 
         """
         if project_path(path) is None:
@@ -151,8 +151,8 @@ class DirectoryStructure(object):
         self.__set_dir = "settings"
         self.__sym_dir = "symmetry"
         self.__clex_dir = "cluster_expansions"
-    
-    
+
+
     # ** Query filesystem **
 
     def all_bset(self):
@@ -162,7 +162,7 @@ class DirectoryStructure(object):
     def all_calctype(self):
       """Check filesystem directory structure and return list of all calctype names"""
       return self.__all_settings("calctype", join(self.path, self.__calc_dir, self.__set_dir))
-    
+
     def all_ref(self, calctype):
       """Check filesystem directory structure and return list of all ref names for a given calctype"""
       return self.__all_settings("ref", self.calc_settings_dir(calctype))
@@ -240,7 +240,7 @@ class DirectoryStructure(object):
     def clust(self, clex):
       """Returns path to the clust.json file"""
       return join(self.bset_dir(clex), "clust.json")
-    
+
     def basis(self, clex):
       """Returns path to the basis.json file"""
       return join(self.bset_dir(clex), "basis.json")
@@ -256,7 +256,7 @@ class DirectoryStructure(object):
     def clexulator_o(self, project, clex):
       """Returns path to global clexulator.o file"""
       return join(self.bset_dir(clex), (project + "_Clexulator.o"))
-    
+
     def clexulator_so(self, project, clex):
       """Returns path to global clexulator so file"""
       return join(self.bset_dir(clex), (project + "_Clexulator.so"))
@@ -266,50 +266,50 @@ class DirectoryStructure(object):
 
     def settings_path_crawl(self, filename, configname, clex):
         """
-        Returns the path to the first file named 'filename' found in the settings 
+        Returns the path to the first file named 'filename' found in the settings
         directories.
-        
+
         Searches:
           1) self.configuration_calc_settings_dir(configname, clex)
           2) self.supercell_calc_settings_dir(scelname, clex)
           3) self.calc_settings_dir(clex)
-          DirectoryStructure.configuration_calc_settings_dir(configname, clex)Crawl casm directory structure starting at configdir and moving upwards 
-        
+          DirectoryStructure.configuration_calc_settings_dir(configname, clex)Crawl casm directory structure starting at configdir and moving upwards
+
         Returns None if file named 'filename' not found in any of the three directories.
-        
-        
+
+
         Arguments
         ---------
           filename: str
             The name of the file being searched for
-          
+
           configname: str
             The name of the configuration
-            
+
           clex: a casm.project.ClexDescription instance
             Used to specify the calctype to find settings for
-          
-        
+
+
         Returns
         ---------
           filepath: str or None
             The path to the first file named 'filename' found in the settings
             directories, or None if not found.
-      
+
         """
         filepath = join(self.configuration_calc_settings_dir(configname, clex), filename)
         if os.path.exists(filepath):
           return filepath
-        
+
         scelname = configname.split('/')[0]
         filepath = join(self.supercell_calc_settings_dir(scelname, clex), filename)
         if os.path.exists(filepath):
           return filepath
-        
+
         filepath = join(self.calc_settings_dir(clex), filename)
         if os.path.exists(filepath):
           return filepath
-        
+
         return None
 
     def supercell_dir(self, scelname):
@@ -319,11 +319,11 @@ class DirectoryStructure(object):
     def configuration_dir(self, configname):
       """Return configuration directory path (configname has format SCELV_A_B_C_D_E_F/I)"""
       return join(self.path, self.__calc_dir, configname)
-    
+
     def POS(self, configname):
       """Return path to POS file"""
       return join(self.configuration_dir(configname), "POS")
-    
+
     def calctype_dir(self, configname, clex):
       """Return calctype directory path (e.g. training_data/SCEL_...../0/calctype.default"""
       return join(self.configuration_dir(configname),self.__calctype(clex.calctype))
@@ -352,11 +352,11 @@ class DirectoryStructure(object):
     def composition_axes(self):
       """Return composition axes file path"""
       return join(self.casm_dir(), "composition_axes.json")
-    
+
     def chemical_reference(self, clex):
       """Return chemical reference file path"""
       return join(self.ref_dir(clex), "chemical_reference.json")
-    
+
 
     # -- Cluster expansions --------
 
@@ -367,12 +367,12 @@ class DirectoryStructure(object):
     def eci_dir(self, clex):
       """
       Returns path to eci directory
-      
+
       Arguments
       ---------
         clex: a casm.project.ClexDescription instance
           Specifies the cluster expansion to get the eci directory for
-        
+
       Returns
       -------
         p: str
@@ -383,12 +383,12 @@ class DirectoryStructure(object):
     def eci(self, clex):
       """
       Returns path to eci.json
-      
+
       Arguments
       ---------
         clex: a casm.project.ClexDescription instance
           Specifies the cluster expansion to get the eci.json for
-        
+
       Returns
       -------
         p: str
@@ -414,61 +414,61 @@ class DirectoryStructure(object):
     def __eci(self, eci):
       return "eci." + eci
 
-    
+
     def __all_settings(self, pattern, location):
       """
       Find all directories at 'location' that match 'pattern.something'
       and return a std::vector of the 'something'
       """
-      
+
       all = [];
       pattern += ".";
 
       # get all
       if not os.path.exists(location):
         return all
-      
+
       for item in os.listdir(location):
         if os.path.isdir(os.path.join(location,item)) and item[:len(pattern)] == pattern:
           all.append(item[len(pattern):])
       return sorted(all)
 
-    
+
 
 class Project(object):
     """The Project class contains information about a CASM project
-    
+
     Attributes
     ----------
-      
+
       path: str
         Path to project root directory
-      
+
       name: str
         Project name
-      
+
       settings: casm.project.ProjectSettings instance
         Contains project settings
-      
+
       dir: casm.project.DirectoryStructure instance
         Provides file and directory locations within the project
-      
+
       prim: casm.project.Prim instance
         Represents the primitive crystal structure
-      
+
       composition_axes: casm.project.CompositionAxes
         Currently selected composition axes
-      
+
       all_composition_axes: dict(str:casm.project.CompositionAxes)
         Dict containing name:CompositionAxes pairs, including both standard and
         custom composition axes
-      
+
       casm_exe: str
         The casm CLI executable to use when necessary
-      
+
       verbose: bool
         How much to print to stdout
-      
+
     """
     def __init__(self, path=None, casm_exe=None, verbose=True):
       """
@@ -476,45 +476,45 @@ class Project(object):
 
       Arguments
       ----------
-        
-        path: str, optional, default=None 
-          Path to project root directory. Default=None uses project containing 
+
+        path: str, optional, default=None
+          Path to project root directory. Default=None uses project containing
           current working directory
-        
+
         casm_exe: str, optional, default=None
           CASM executable to use for command line interface. Default
           uses $CASM if it exists in the environment, else "casm".
-        
+
         verbose: bool, optional, default=True
           How much to print to stdout
-      
+
       """
-      
+
       # will hold a ctypes.c_void_p when loading CASM project into memory
       self._ptr = None
-      
+
       # will keep a casm.API instance
       self._api = None
-      
+
       # set path to this CASM project
       if project_path(path) is None:
         if path is None:
           raise Exception("No CASM project found using " + os.getcwd())
         else:
           raise Exception("No CASM project found using " + path)
-      
+
       # set executable name
       if casm_exe is None:
         if "CASM" in os.environ:
           casm_exe = os.environ["CASM"]
         else:
           casm_exe = "casm"
-      
+
       self.path = project_path(path)
       self.__refresh()
       self.casm_exe = casm_exe
       self.verbose = verbose
-      
+
       self.all_composition_axes = {}
       if os.path.exists(self.dir.composition_axes()):
           with open(self.dir.composition_axes(), 'r') as f:
@@ -526,12 +526,12 @@ class Project(object):
                   for key, val in six.iteritems(data["custom_axes"]):
                       self.all_composition_axes[key] = CompositionAxes(key, val)
               self.composition_axes = self.all_composition_axes[data["current_axes"]]
-    
-    
+
+
     def __del__(self):
       self.__unload()
-    
-    
+
+
     def __load(self):
       """
       Explicitly load CASM project into memory.
@@ -542,15 +542,15 @@ class Project(object):
           streamptr = self._api.stdout()
         else:
           streamptr = self._api.nullstream()
-        
+
         if self.verbose:
           errstreamptr = self._api.stderr()
         else:
           errstreamptr = self._api.nullstream()
-        
+
         self._ptr = self._api.primclex_new(self.path, streamptr, streamptr, errstreamptr)
-        
-    
+
+
     def __unload(self):
       """
       Explicitly unload CASM project from memory.
@@ -558,29 +558,29 @@ class Project(object):
       if self._ptr is not None:
         self._api.primclex_delete(self._ptr)
         self._ptr = None
-    
-    
+
+
     def __refresh(self):
       """
       Reload self.settings and self.dir
-      
+
       Use this after adding or modifying files in the CASM project but no
       special call to refresh PrimClex properties is required
       """
       self.dir = DirectoryStructure(self.path)
       self.settings = ProjectSettings(self.path)
       self._prim = None
-    
+
     @property
     def prim(self):
         if self._prim is None:
             self._prim = Prim(self)
         return self._prim
-    
+
     @property
     def name(self):
         return self.settings.data['name']
-    
+
     def refresh(self, read_settings=False, read_composition=False, read_chem_ref=False, read_configs=False, clear_clex=False):
       """
       Refresh PrimClex properties to reflect changes to CASM project files.
@@ -589,161 +589,161 @@ class Project(object):
         self.__refresh()
       if self._ptr is not None:
         self._api.primclex_refresh(
-          self.data(), 
+          self.data(),
           read_settings,
           read_composition,
           read_chem_ref,
           read_configs,
           clear_clex)
-      
-    
+
+
     def data(self):
       """
       Returns a 'ctypes.c_void_p' that points to a CASM project. (PrimClex)
       """
       self.__load()
       return self._ptr
-    
-    
+
+
     def command(self, args):
       """
-      Execute a command via the c api. 
-      
+      Execute a command via the c api.
+
       Args:
         args: A string containing the command to be executed. Ex: "select --set-on -o /abspath/to/my_selection"
-      
+
       Returns:
         (stdout, stderr, returncode): The result of running the command via the command line iterface
       """
       return self._command_via_capi(args)
-    
-        
+
+
     def _command_via_capi(self, args):
       """
-      Execute a command via the c api. 
-      
+      Execute a command via the c api.
+
       Args:
         args: A string containing the command to be executed. Ex: "select --set-on -o /abspath/to/my_selection"
-      
+
       Returns:
-        (stdout, stderr, returncode): The result of running the command via the 
+        (stdout, stderr, returncode): The result of running the command via the
             command line iterface. 'stdout' and 'stderr' are in text type ('unicode'/'str').
-            
+
       """
       # this also ensures self._api is not None
       data = self.data()
-      
+
       # construct stringstream objects to capture stdout, debug, stderr
       ss = self._api.ostringstream_new()
       ss_debug = self._api.ostringstream_new()
       ss_err = self._api.ostringstream_new()
-      
+
       res = self._api(args, self.data(), self.path, ss, ss_debug, ss_err)
-      
+
       # copy strings and delete stringstreams
       stdout = self._api.ostringstream_to_str(ss)
       self._api.ostringstream_delete(ss)
-      
+
       self._api.ostringstream_delete(ss_debug)
-      
+
       stderr = self._api.ostringstream_to_str(ss_err)
       self._api.ostringstream_delete(ss_err)
-      
+
       self.__refresh()
       return (stdout.decode('utf-8'), stderr.decode('utf-8'), res)
 
 
 class Prim(object):
     """The Primitive Crystal Structure
-    
+
     Attributes
     ----------
-    
+
         proj: casm.Project
           the CASM project the Prim belongs to
-    
+
         lattice_matrix: np.array of shape (3, 3)
           lattice vectors as matrix columns
-        
+
         lattice_parameters: dict
           Lattice parameters and angles (in degrees), as:
             {'a':a, 'b':b, 'c':c, 'alpha':alpha, 'beta':beta, 'gamma':gamma}
-        
+
         basis: List(dict)
           crystal basis, as read directly from prim.json (format may change)
-        
+
         coordinate_mode: str
           crystal basis coordinate_mode, as read directly from prim.json (format
           may change)
-        
+
         lattice_symmetry_s: str
           lattice point group, in Schoenflies notation
-        
+
         lattice_symmetry_hm: str
           lattice point group, in Hermann-Mauguin notation
-        
+
         lattice_system: str
           lattice system name, ('cubic', 'hexagonal', 'rhombohedral', etc.)
-        
+
         crystal_symmetry_s: str
           crystal point group, in Schoenflies notation
-    
+
         crystal_symmetry_hm: str
           crystal point group, in Hermann-Mauguin notation
-    
+
         crystal_system: str
           crystal system name, ('cubic', 'hexagonal', 'trigonal', etc.)
-        
+
         crystal_family: str
           crystal family name, ('cubic', 'hexagonal', etc.)
-        
+
         space_group_number: str
           range of possible space group number
-        
+
         components: List[str]
           occupational components
-        
+
         elements: List[str]
           all allowed elements
-        
+
         n_independent_compositions: int
           number of independent composition axes
-        
+
         degrees_of_freedom: List[str]
           allowed degrees of freedom, from:
             'occupation'
-    
+
     """
-    
+
     def __init__(self, proj):
         """
         Construct a CASM Prim
 
         Arguments
         ---------
-          
+
           proj: casm.Project, optional, default=Project containing the current working directory
             the CASM project the Prim belongs to
-        
+
         """
         if proj == None:
             proj = Project()
         elif not isinstance(proj, Project):
             raise Exception("Error constructing Prim: proj argument is not a CASM Project")
         self.proj = proj
-        
-        # raw prim.json (for some properties not yet supported in the casm.project API)
+
+        # raw prim.json (for some properties not yet supported in the API)
         with open(self.proj.dir.prim()) as f:
             raw_prim = json.load(f)
         self.lattice_matrix = np.array(raw_prim['lattice_vectors']).transpose()
         self.basis = raw_prim['basis']
         self.coordinate_mode = raw_prim['coordinate_mode']
-    
+
         def _angle(a, b):
             return math.degrees(math.acos(
                 np.dot(a,b) / (np.linalg.norm(a) * np.linalg.norm(b))
             ))
-    
+
         def _lattice_parameters(L):
             a = np.linalg.norm(L[:,0])
             b = np.linalg.norm(L[:,1])
@@ -753,25 +753,25 @@ class Prim(object):
             gamma = _angle(L[:,0], L[:,1])
             return {'a':a, 'b':b, 'c':c, 'alpha':alpha, 'beta':beta, 'gamma':gamma}
         self.lattice_parameters = _lattice_parameters(self.lattice_matrix)
-        
+
         (stdout, stderr, returncode) = proj.command("sym")
-        
+
         # lattice symmetry
         self.lattice_symmetry_s = syminfo.lattice_symmetry(stdout)
         self.lattice_symmetry_hm = syminfo.hm_symmetry(self.lattice_symmetry_s)
         self.lattice_system = syminfo.lattice_system(self.lattice_symmetry_s)
-    
+
         # crystal symmetry
         self.crystal_symmetry_s = syminfo.crystal_symmetry(stdout)
         self.crystal_symmetry_hm = syminfo.hm_symmetry(self.crystal_symmetry_s)
         self.crystal_system = syminfo.crystal_system(self.crystal_symmetry_s)
         self.crystal_family = syminfo.crystal_family(self.crystal_symmetry_s)
         self.space_group_number = syminfo.space_group_number_map[self.crystal_symmetry_s]
-    
+
         # composition (v0.2.X: elements and components are identical, only 'occupation' allowed)
         with open(self.proj.dir.composition_axes()) as f:
             raw_composition_axes = json.load(f)
-        
+
         self.components = raw_composition_axes['standard_axes']['0']['components']
         self.elements = self.components
         self.n_independent_compositions = raw_composition_axes['standard_axes']['0']['independent_compositions']
@@ -780,36 +780,36 @@ class Prim(object):
 
 class CompositionAxes(object):
     """A composition axes object
-    
+
     Attributes
     ----------
-    
+
         name: str
           composition axes name
-        
+
         components: List[str]
           occupational components
-        
+
         n_independent_compositions: int
           number of independent composition axes
-        
+
         mol_formula: str
           number of each component in terms of the parametric compositions
-        
+
         param_formula: str
           parametric compositions in terms of the number of components
-        
+
         end_members: dict of np.array of shape=(n_components,)
           the number of components per unit cell in each end member state, in form:
           {'origin':np.array, 'a':np.array, 'b', np.array, ...}. Order matches
           that given by self.components.
-          
-    
+
+
     """
     def __init__(self, name, data):
         self._name = name
         self._data = data
-        
+
         self._end_members = {}
         for c in ascii_lowercase:
             if c in self._data:
@@ -821,32 +821,27 @@ class CompositionAxes(object):
     @property
     def name(self):
         return self._name
-    
+
     @property
     def components(self):
         return self._data['components']
-    
+
     @property
     def n_independent_compositions(self):
         return self._data['independent_compositions']
-    
+
     @property
     def mol_formula(self):
         return self._data['mol_formula']
-    
+
     @property
     def param_formula(self):
         return self._data['param_formula']
-    
+
     @property
     def origin(self):
         return self._origin
-    
+
     @property
     def end_members(self):
         return self._end_members
-    
-    
-    
-    
-
