+++ python/casm/casm/scripts/casm.py
@@ -1,8 +1,10 @@
 from __future__ import (absolute_import, division, print_function, unicode_literals)
 from builtins import *
 
-import warnings
-warnings.filterwarnings("ignore", category=DeprecationWarning)
+# import warnings
+# warnings.filterwarnings("ignore", category=DeprecationWarning)
+# warnings.filterwarnings("ignore", message="numpy.dtype size changed")
+# warnings.filterwarnings("ignore", message="numpy.ufunc size changed")
 
 import argparse
 import json
@@ -14,7 +16,7 @@ import sys
 import casm.scripts.casm_calc
 import casm.scripts.casm_learn
 import casm.scripts.casm_plot
-from casm.project import API, Project
+from casm.api import API
 
 def _exec(argv=None):
     subprocess.Popen(argv)
@@ -24,7 +26,7 @@ def _shell(argv=None):
     os.environ['PS1'] = "<shell> " + os.environ['PS1']
     pty.spawn('/bin/bash')
     os.environ['PS1'] = orig
-        
+
 
 def main(argv=None):
     if argv is None:
@@ -36,29 +38,29 @@ def main(argv=None):
     except:
         libcasm_commands = []
         found_libcasm = False
-    
+
     python_commands = {
         'calc': casm.scripts.casm_calc.main,
         'learn': casm.scripts.casm_learn.main,
         'plot': casm.scripts.casm_plot.main,
         'shell': _shell,
         'exec': _exec}
-    
+
     commands = sorted(libcasm_commands + list(python_commands.keys()))
-    
+
     if not found_libcasm:
         print("Could not find libcasm. Please check your installation.")
         return
-    
+
     parser = argparse.ArgumentParser(description = 'CASM: First-principles based statistical mechanics')
     parser.add_argument('command', help="CASM command to execute", type=str, default="", nargs="?", metavar="<command>")
     parser.add_argument('args', help="CASM command arguments", nargs=argparse.REMAINDER, metavar="...")
     parser.add_argument('--desc', help="Print command list", action="store_true", default=False)
     parser.add_argument('--version', help="Print casm version", action="store_true", default=False)
     parser.add_argument('--path', help="Path to project. Default uses project containing current working directory.", type=str, default="")
-    
+
     args = parser.parse_args(argv)
-    
+
     if args.version:
         args.command = "version"
         args.args = []
